test: move api-proxy suite
