+++ internal/jujuapi/websocket_test.go
@@ -11,12 +11,9 @@ import (
 	"net/http"
 	"net/http/httptest"
 	"net/url"
-	"strings"
 
 	"github.com/go-chi/chi/v5"
-	"github.com/juju/errors"
 	"github.com/juju/juju/api"
-	"github.com/juju/juju/api/client/client"
 	"github.com/juju/juju/rpc/jsoncodec"
 	jujuparams "github.com/juju/juju/rpc/params"
 	"github.com/juju/names/v5"
@@ -172,160 +169,6 @@ func (s *websocketSuite) open(c *gc.C, info *api.Info, username string) api.Conn
 	c.Assert(err, gc.Equals, nil)
 	return conn
 }
-
-func (s *websocketSuite) openCustomLoginProvider(c *gc.C, info *api.Info, username string, lp api.LoginProvider) (api.Connection, error) {
-	ld := loginDetails{info: info, username: username, lp: lp}
-	return s.openNoAssert(c, ld)
-}
-
-type apiProxySuite struct {
-	websocketSuite
-}
-
-var _ = gc.Suite(&apiProxySuite{})
-
-func (s *apiProxySuite) TestConnectToModel(c *gc.C) {
-	conn := s.open(c, &api.Info{
-		ModelTag:  s.Model.ResourceTag(),
-		SkipLogin: true,
-	}, "test")
-	defer conn.Close()
-	var resp map[string]interface{}
-	err := conn.APICall("Admin", 3, "", "TestMethod", nil, &resp)
-	c.Assert(err, gc.ErrorMatches, `(?s).*no such request - method Admin.TestMethod is not implemented \(not implemented\).*`)
-}
-
-// TestSessionTokenLoginProvider verifies that the session token login provider works as expected.
-// We do this by using a mock authenticator that simulates polling an OIDC server and verifying that
-// the user would be prompted with a login URL and fake the user login via the `EnableDeviceFlow` method.
-func (s *apiProxySuite) TestSessionTokenLoginProvider(c *gc.C) {
-	ctx := context.Background()
-	alice := names.NewUserTag("alice@canonical.com")
-	aliceUser := openfga.NewUser(&dbmodel.Identity{Name: alice.Id()}, s.JIMM.OpenFGAClient)
-	err := aliceUser.SetControllerAccess(ctx, s.Model.Controller.ResourceTag(), ofganames.AdministratorRelation)
-	c.Assert(err, gc.IsNil)
-	var output bytes.Buffer
-	s.EnableDeviceFlow(aliceUser.Name)
-	conn, err := s.openCustomLoginProvider(c, &api.Info{
-		ModelTag:  s.Model.ResourceTag(),
-		SkipLogin: false,
-	}, "alice", api.NewSessionTokenLoginProvider("", &output, func(s string) {}))
-	c.Assert(err, gc.IsNil)
-	defer conn.Close()
-	c.Check(err, gc.Equals, nil)
-	outputNoNewLine := strings.ReplaceAll(output.String(), "\n", "")
-	c.Check(outputNoNewLine, gc.Matches, `Please visit .* and enter code.*`)
-}
-
-// TestAgentLoginReturnsRedirect verifies that the agent login returns a redirect error
-// when trying to connect to the model proxy. We don't use the helper methods
-// for opening the connection because we want to test authenticating as an agent.
-func (s *apiProxySuite) TestAgentLoginReturnsRedirect(c *gc.C) {
-	u, err := url.Parse(s.HTTP.URL)
-	c.Assert(err, gc.Equals, nil)
-
-	info := api.Info{
-		Tag:      names.NewUnitTag("ubuntu/1"),
-		ModelTag: s.Model.ResourceTag(),
-		Addrs:    []string{u.Host},
-	}
-	dialOpts := api.DialOpts{
-		InsecureSkipVerify: true,
-	}
-	_, err = api.Open(&info, dialOpts)
-	c.Assert(err, gc.NotNil)
-	redirectErr, ok := errors.Cause(err).(*api.RedirectError)
-	c.Check(ok, gc.Equals, true)
-	c.Assert(redirectErr.Servers, gc.HasLen, 1)
-	servers := redirectErr.Servers[0].HostPorts().Strings()
-	c.Assert(servers, gc.HasLen, 1)
-	c.Check(servers[0], gc.Matches, "localhost:[0-9]+")
-	c.Check(len(redirectErr.CACert), gc.Not(gc.Equals), 0)
-}
-
-func (s *apiProxySuite) TestAgentLoginModelDoesNotExist(c *gc.C) {
-	u, err := url.Parse(s.HTTP.URL)
-	c.Assert(err, gc.Equals, nil)
-
-	info := api.Info{
-		Tag:      names.NewUnitTag("ubuntu/1"),
-		ModelTag: names.NewModelTag("00000000-0000-0000-0000-000000000000"),
-		Addrs:    []string{u.Host},
-	}
-	dialOpts := api.DialOpts{
-		InsecureSkipVerify: true,
-	}
-	_, err = api.Open(&info, dialOpts)
-	c.Assert(err, gc.NotNil)
-	c.Check(err.Error(), gc.Matches, `failed to find model: model not found.*`)
-}
-
-type logger struct{}
-
-func (l logger) Errorf(string, ...interface{}) {}
-
-func (s *apiProxySuite) TestModelStatus(c *gc.C) {
-	conn := s.open(c, &api.Info{
-		ModelTag:  s.Model.ResourceTag(),
-		SkipLogin: false,
-	}, "alice@canonical.com")
-	defer conn.Close()
-	jujuClient := client.NewClient(conn, logger{})
-	status, err := jujuClient.Status(nil)
-	c.Check(err, gc.IsNil)
-	c.Check(status, gc.Not(gc.IsNil))
-	c.Check(status.Model.Name, gc.Equals, s.Model.Name)
-}
-
-func (s *apiProxySuite) TestModelStatusWithoutPermission(c *gc.C) {
-	fooUser := openfga.NewUser(&dbmodel.Identity{Name: "foo@canonical.com"}, s.JIMM.OpenFGAClient)
-	var output bytes.Buffer
-	s.EnableDeviceFlow(fooUser.Name)
-	conn, err := s.openCustomLoginProvider(c, &api.Info{
-		ModelTag:  s.Model.ResourceTag(),
-		SkipLogin: false,
-	}, "foo", api.NewSessionTokenLoginProvider("", &output, func(s string) {}))
-	c.Check(err, gc.ErrorMatches, "permission denied .*")
-	if conn != nil {
-		defer conn.Close()
-	}
-	outputNoNewLine := strings.ReplaceAll(output.String(), "\n", "")
-	c.Check(outputNoNewLine, gc.Matches, `Please visit .* and enter code.*`)
-}
-
-// TODO(Kian): This test aims to verify that JIMM gracefully handles clients that end their connection
-// during the login flow after JIMM starts polling the OIDC server.
-// After https://github.com/juju/juju/pull/17606 lands we can begin work on this.
-// The API connection's login method should be refactored use the login provider stored on the state struct.
-// func (s *apiProxySuite) TestDeviceFlowCancelDuringPolling(c *gc.C) {
-// 	ctx := context.Background()
-// 	alice := names.NewUserTag("alice@canonical.com")
-// 	aliceUser := openfga.NewUser(&dbmodel.Identity{Name: alice.Id()}, s.JIMM.OpenFGAClient)
-// 	err := aliceUser.SetControllerAccess(ctx, s.Model.Controller.ResourceTag(), ofganames.AdministratorRelation)
-// 	c.Assert(err, gc.IsNil)
-// 	var cliOutput string
-// 	_ = cliOutput
-// 	outputFunc := func(format string, a ...any) error {
-// 		cliOutput = fmt.Sprintf(format, a)
-// 		return nil
-// 	}
-// 	var wg sync.WaitGroup
-// 	var conn api.Connection
-// 	wg.Add(1)
-// 	go func() {
-// 		defer wg.Done()
-// 		conn, err = s.openCustomLP(c, &api.Info{
-// 			ModelTag:  s.Model.ResourceTag(),
-// 			SkipLogin: true,
-// 		}, "alice", api.NewSessionTokenLoginProvider("", outputFunc, func(s string) error { return nil }))
-// 		c.Assert(err, gc.IsNil)
-// 	}()
-// 	conn.Login()
-//  // Close the connection after the cliOutput is filled.
-// 	c.Assert(err, gc.Equals, nil)
-// }
-
-// TODO(CSS-7331) Add more tests for model proxy and new login methods.
 
 type pathTestSuite struct{}
